refactor: remove factory and simplify itemEnum
