+++ Java/src/main/java/com/gildedrose/GildedRose.java
@@ -1,8 +1,5 @@
 package com.gildedrose;
 
-import com.gildedrose.factory.Factory;
-import com.gildedrose.factory.ItemEnum;
-
 class GildedRose {
     Item[] items;
 
@@ -11,19 +8,9 @@ class GildedRose {
     }
 
     public void updateQuality() {
-        Factory factory = new Factory();
         for (Item item : items) {
-            factory.setItemEnum(converItemNameToItemEnum(item.name));
-            factory.updateItem(item);
+            ItemEnum itemEnum = ItemEnum.getItemEnum(item.name);
+            itemEnum.update(item);
         }
-    }
-
-    private ItemEnum converItemNameToItemEnum(String itemName) {
-        for (ItemEnum itemEnum : ItemEnum.values()) {
-            if (itemEnum.getValue().equals(itemName)) {
-                return itemEnum;
-            }
-        }
-        return ItemEnum.NORMAL;
     }
 }
